requirements.txt
torch==1.8.1
torchvision==0.9.1
opencv_python==4.5.2.54
importlib_metadata==4.5.0
numpy==1.19.5
tqdm==4.61.0
matplotlib==3.4.2
six==1.15.0
pandas==1.2.4
fire==0.4.0
ipdb==0.13.9
ipython==7.27.0
keras==2.6.0
Pillow==8.3.2
pycocotools==2.0.2
PyQt5==5.15.4
pytorch_lightning==1.4.7
scikit_learn==0.24.2
scikit-image==0.19.2
tensorflow==2.6.0
torchnet==0.0.4
visdom==0.1.8.9
